+++ src/main/java/ru/sbrf/payment/client/AccountDeposit.java
@@ -3,7 +3,7 @@ package ru.sbrf.payment.client;
 import ru.sbrf.payment.common.Currency;
 
 public class AccountDeposit extends Account {
-    public AccountDeposit(String accountNumber, Currency currency, float balance) {
+    public AccountDeposit(String accountNumber, Currency currency, long balance) {
         super(accountNumber, currency, balance);
     }
 }
